Move icons into a single module
