--- src/icons/edit-icon.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react'
-import IconFrame from './icon-frame'
-
-const EditIcon = () => (
-  <IconFrame>
-    <path
-      opacity="0.54"
-      d="M2,12.88V16h3.12L14,7.12L10.88,4L2,12.88z M16.76,4.37c0.33-0.33,0.33-0.85,0-1.18l-1.949-1.95c-0.33-0.33-0.851-0.33-1.181,0L12,2.88L15.12,6L16.76,4.37z"
-    />
-  </IconFrame>
-)
-
-export default EditIcon
