--- src/icons/gallery-icon.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react'
-import IconFrame from './icon-frame'
-
-const GalleryIcon = () => (
-  <IconFrame>
-    <path d="M4 11h5V5H4v6zm0 7h5v-6H4v6zm6 0h5v-6h-5v6zm6 0h5v-6h-5v6zm-6-7h5V5h-5v6zm6-6v6h5V5h-5z" />
-    <path d="M0 0h24v24H0z" fill="none" />
-  </IconFrame>
-)
-
-export default GalleryIcon
