+++ src/icons/icon-frame.tsx
@@ -6,17 +6,15 @@ interface Props {
   width?: number
 }
 
-const IconFrame: React.SFC<Props> = ({ width, height, fill, children }) => {
-  const viewBox = `0 0 ${width} ${height}`
+const IconFrame: React.SFC<Props> = ({ children, ...props }) => {
+  const viewBox = `0 0 ${props.width} ${props.height}`
 
   return (
     <svg
       className="ars-icon"
-      fill={fill}
-      height={height}
-      width={width}
-      viewBox={viewBox}
       xmlns="http://www.w3.org/2000/svg"
+      viewBox={viewBox}
+      {...props}
     >
       {children}
     </svg>
@@ -27,6 +25,13 @@ IconFrame.defaultProps = {
   fill: '#000000',
   height: 24,
   width: 24
+}
+
+export function generateIcon<IconProps>(
+  callback: (props: IconProps) => React.ReactNode
+): React.SFC<IconProps> {
+  return (props: IconProps) =>
+    React.createElement(IconFrame, props, callback(props))
 }
 
 export default IconFrame
