+++ src/icons/index.js
@@ -1,0 +1,46 @@
+import * as React from 'react'
+import { generateIcon } from './icon-frame'
+
+export const ClearIcon = generateIcon(() => (
+  <>
+    <title>Clear</title>
+    <path d="M0 0h24v24H0z" fill="none" />
+    <path d="M8 4v1.45l2 2V4h4v4h-3.45l2 2H14v1.45l2 2V10h4v4h-3.45l2 2H20v1.45l2 2V4c0-1.1-.9-2-2-2H4.55l2 2H8zm8 0h4v4h-4V4zM1.27 1.27L0 2.55l2 2V20c0 1.1.9 2 2 2h15.46l2 2 1.27-1.27L1.27 1.27zM10 12.55L11.45 14H10v-1.45zm-6-6L5.45 8H4V6.55zM8 20H4v-4h4v4zm0-6H4v-4h3.45l.55.55V14zm6 6h-4v-4h3.45l.55.54V20zm2 0v-1.46L17.46 20H16z" />
+  </>
+))
+
+export const EditIcon = generateIcon(() => (
+  <path
+    opacity="0.54"
+    d="M2,12.88V16h3.12L14,7.12L10.88,4L2,12.88z M16.76,4.37c0.33-0.33,0.33-0.85,0-1.18l-1.949-1.95c-0.33-0.33-0.851-0.33-1.181,0L12,2.88L15.12,6L16.76,4.37z"
+  />
+))
+
+export const GalleryIcon = generateIcon(() => (
+  <>
+    <path d="M4 11h5V5H4v6zm0 7h5v-6H4v6zm6 0h5v-6h-5v6zm6 0h5v-6h-5v6zm-6-7h5V5h-5v6zm6-6v6h5V5h-5z" />
+    <path d="M0 0h24v24H0z" fill="none" />
+  </>
+))
+
+export const RefreshIcon = generateIcon(() => (
+  <path d="M17.65 6.35C16.2 4.9 14.21 4 12 4c-4.42 0-7.99 3.58-7.99 8s3.57 8 7.99 8c3.73 0 6.84-2.55 7.73-6h-2.08c-.82 2.33-3.04 4-5.65 4-3.31 0-6-2.69-6-6s2.69-6 6-6c1.66 0 3.14.69 4.22 1.78L13 11h7V4l-2.35 2.35z" />
+))
+
+export const SearchIcon = generateIcon(() => (
+  <>
+    <title>Enter Search Term</title>
+    <path
+      opacity="0.54"
+      d="M15.5,14h-0.79l-0.28-0.27C15.41,12.59,16,11.11,16,9.5C16,5.91,13.09,3,9.5,3S3,5.91,3,9.5S5.91,16,9.5,16c1.61,0,3.09-0.59,4.23-1.57L14,14.71v0.79l5,4.99L20.49,19L15.5,14z M9.5,14C7.01,14,5,11.99,5,9.5S7.01,5,9.5,5S14,7.01,14,9.5S11.99,14,9.5,14z"
+    />
+    <path fill="none" d="M0,0h24v24H0V0z" />
+  </>
+))
+
+export const TableIcon = generateIcon(() => (
+  <>
+    <path d="M3 13h2v-2H3v2zm0 4h2v-2H3v2zm0-8h2V7H3v2zm4 4h14v-2H7v2zm0 4h14v-2H7v2zM7 7v2h14V7H7z" />
+    <path d="M0 0h24v24H0z" fill="none" />
+  </>
+))
